// front/src/types/UserTypes.tsx
export interface UserProps {
  _id: string;
  uniqueId: string;
  username: string;
  name: string;
  avatar: string[];
  email: string;
  mobile_no: string;
  alt_mobile_no: string;
  address: string;
  pincode: string;
  city: string;
  state: string;
  country: string;
  isProfessional: boolean;
  verified: boolean;
  role: string;
}